// src/app/models/task.interface.ts
export interface TaskI {
    id?: string;
    codigo: string;
    nombre: string;
    categoria: string;
    descripcion: string;
    tallas: string;
    precio: number;
    preciodolar: number;
    img: string;
}
export interface TaskIGaleriasGenerales {
    id?: string;
    codigo: string;
    nombre: string;
    categoria: string;
    descripcion: string;
    tallas: string;
    precio: number;
    preciodolar: number;
    img: string;
}